+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ```
 v0.0.2:
-  date: TBD
+  date: 2016-01-29
   changes:
     - Created IDEAS.md
+    - Created friendly.c
 
 v0.0.1:
   date: 2016-01-28